Fix starting index bug
The "index" key of the returned dict is expected to have the index of
the first item on the returned data in the dataset.
In previous implementation, "index" was just the provided parameter
unaltered. This introduces a bug if the element at that index in the
dataset was deleted prior to retrieval.